feat: 보너스볼 중복 제외 전부 구현, 테케 통과

+++ src/main/java/lotto/domain/IncomeRate.java
@@ -18,7 +18,7 @@ public class IncomeRate {
         return this.investment;
     }
 
-    public int getIncomeRate() {
-        return this.income / this.investment * 100;
+    public double getIncomeRate() {
+        return (double) this.income / this.investment * 100;
     }
 }
